fix(Dashboard): Update line and table widget threashold 'to' and 'from' feilds data type from int to float
Updated the data type for the "to" and "From" feilds in "Threashold" in line and table widget from int to float(GitHub #2703)

+++ newrelic/structures_newrelic_one_dashboard.go
@@ -504,11 +504,11 @@ func expandDashboardLineWidgetConfigurationThresholdInput(d *schema.ResourceData
 				// and assign them to respective attributes of the DashboardLineWidgetThresholdThresholdInput object
 
 				if v, ok := lineWidgetThresholdInInputSingularInterface["from"]; ok {
-					t := v.(int)
+					t := v.(float64)
 					lineWidgetThresholdToBeAdded.From = &t
 				}
 				if v, ok := lineWidgetThresholdInInputSingularInterface["to"]; ok {
-					t := v.(int)
+					t := v.(float64)
 					lineWidgetThresholdToBeAdded.To = &t
 				}
 				if v, ok := lineWidgetThresholdInInputSingularInterface["name"]; ok {
@@ -559,11 +559,11 @@ func expandDashboardTableWidgetConfigurationThresholdInput(d *schema.ResourceDat
 				// and assign them to respective attributes of the DashboardTableWidgetThresholdInput object
 
 				if v, ok := tableWidgetThresholdInInputSingularInterface["from"]; ok {
-					t := v.(int)
+					t := v.(float64)
 					tableWidgetThresholdToBeAdded.From = &t
 				}
 				if v, ok := tableWidgetThresholdInInputSingularInterface["to"]; ok {
-					t := v.(int)
+					t := v.(float64)
 					tableWidgetThresholdToBeAdded.To = &t
 				}
 				if v, ok := tableWidgetThresholdInInputSingularInterface["column_name"]; ok {
